Move nano container to own package
